Moving the resolver functions to the core package
